refactor(api): move isSameBinary in test-utils/file.js
