+++ api/tests/prescription/target-profile/integration/application/presenter/pdf/learning-content-pdf-presenter_test.js
@@ -4,8 +4,7 @@ import * as url from 'node:url';
 import pdfLibUtils from 'pdf-lib/cjs/utils/index.js';
 
 import * as learningContentPDFPresenter from '../../../../../../../src/prescription/target-profile/application/presenter/pdf/learning-content-pdf-presenter.js';
-import { domainBuilder, expect, MockDate, sinon } from '../../../../../../test-helper.js';
-import { isSameBinary } from '../../../../../../tooling/binary-comparator.js';
+import { domainBuilder, expect, isSameBinary, MockDate, sinon } from '../../../../../../test-helper.js';
 
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
