+++ api/tests/profile/integration/infrastructure/serializers/pdf/pdf-with-form-serializer_test.js
@@ -5,8 +5,7 @@ import * as url from 'node:url';
 import { PDFDocument } from 'pdf-lib';
 
 import { serializeStream } from '../../../../../../src/profile/infrastructure/serializers/pdf/pdf-with-form-serializer.js';
-import { expect } from '../../../../../test-helper.js';
-import { isSameBinary } from '../../../../../tooling/binary-comparator.js';
+import { expect, isSameBinary } from '../../../../../test-helper.js';
 
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
