+++ api/tests/tooling/test-utils/file.js
@@ -1,0 +1,6 @@
+import fs from 'node:fs/promises';
+
+export async function isSameBinary(referencePath, buffer) {
+  const expectedBuffer = await fs.readFile(referencePath);
+  return expectedBuffer.equals(buffer);
+}
